Move primitive types to their own file
