+++ src/utils.js
@@ -1,0 +1,5 @@
+const isDefined = (e) => e !== undefined && e !== null
+
+module.exports = {
+  isDefined
+}
